+++ packages/explorer/src/app/object/adapters/__tests__/http-client.mock.ts
@@ -2,6 +2,7 @@ import {
   HttpError,
   ParseError,
   type ClientConfig,
+  type DetectionItem,
   type IAlertsClient,
   type listObjectResponse,
   type Newable,
@@ -18,10 +19,20 @@ export const __setTestType = (tt: string) => {
 };
 
 class AlertsClientMock implements IAlertsClient {
+  initClient() {
+    throw new Error("Method not implemented.");
+  }
+  queryDetections<T>(
+    aid: string,
+    parser?: Parser<DetectionItem[], T> | undefined,
+    customModel?: Newable<T> | undefined
+  ): Promise<T> {
+    throw new Error("Method not implemented.");
+  }
   queryObjects<T>(
-    objectFilters: ObjectFilters,
-    parser?: Parser<listObjectResponse, T>,
-    customModel?: Newable<T>
+    _objectFilters: ObjectFilters,
+    _parser?: Parser<listObjectResponse, T>,
+    _customModel?: Newable<T>
   ): Promise<T> {
     return new Promise<T>((resolve, reject) => {
       if (__testType === "success") {
@@ -93,8 +104,18 @@ function isParseError(_: Error) {
   return false;
 }
 
+function serializeParams(params: any, args?: any) {
+  return "serializedParams";
+}
+
+function serializeParamsReverse(queryString: string, args?: any) {
+  return {};
+}
+
 export const mockedModule = {
   AlertsClient,
   isHttpError,
   isParseError,
+  serializeParams,
+  serializeParamsReverse,
 };