+++ packages/explorer/src/app/object/adapters/__tests__/listObjectResponse.mock.ts
@@ -11,6 +11,7 @@ export const objects: PaginatedListEntity<ObjectEntity> = {
   items: [
     {
       aid: "aid123",
+      ndet: 2,
       ra: 111,
       dec: 222,
       firstmjd: 123,