+++ packages/explorer/src/common/test_utils/quasar.ts
@@ -1,13 +1,17 @@
 import { config } from "@vue/test-utils";
+import cloneDeep from "lodash.clonedeep";
 import { createPinia } from "pinia";
 import { afterAll, beforeAll } from "vitest";
 import type { Plugin } from "vue";
+import resetStore from "../reset-store";
 
 export function installPinia() {
-  const globalConfigBackup = config.global;
+  const globalConfigBackup = cloneDeep(config.global);
 
   beforeAll(() => {
-    config.global.plugins.unshift(createPinia() as Plugin);
+    const piniaPlugin = createPinia();
+    piniaPlugin.use(resetStore);
+    config.global.plugins.unshift(piniaPlugin as Plugin);
   });
 
   afterAll(() => {