+++ packages/explorer/src/ui/router/index.ts
@@ -1,41 +1,40 @@
 import { createRouter, createWebHistory } from "vue-router";
-import Search from "@ui/layouts/Search.vue";
-import SearchHome from "@ui/views/SearchHome.vue";
-import SearchResults from "@ui/views/SearchResults.vue";
-import ObjectDetails from "@ui/layouts/ObjectDetails.vue";
-import ObjectDetailsView from "@ui/views/ObjectDetailsView.vue";
+
+export const routes = [
+  {
+    path: "/",
+    name: "search",
+    component: () => import("@ui/layouts/Search.vue"),
+    children: [
+      {
+        path: "/",
+        component: () => import("@ui/views/SearchHome.vue"),
+        name: "search",
+      },
+      {
+        path: "results",
+        name: "results",
+        component: () => import("@ui/views/SearchResults.vue"),
+      },
+    ],
+  },
+  {
+    path: "/object/:objectId",
+    name: "object-details",
+    component: () => import("@ui/layouts/ObjectDetails.vue"),
+    children: [
+      {
+        path: "",
+        component: () => import("@ui/views/ObjectDetailsView.vue"),
+        name: "object",
+      },
+    ],
+  },
+];
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: "/",
-      name: "search",
-      component: Search,
-      children: [
-        {
-          path: "/",
-          component: SearchHome,
-        },
-        {
-          path: "results",
-          component: SearchResults,
-        },
-      ],
-    },
-    {
-      path: "/object/:objectId",
-      name: "object-details",
-      component: ObjectDetails,
-      children: [
-        {
-          path: "",
-          component: ObjectDetailsView,
-          name: "object",
-        },
-      ],
-    },
-  ],
+  routes,
 });
 
 export default router;