+++ packages/hunter/src/App.vue
@@ -1,6 +1,14 @@
 <script setup lang="ts">
+import { onMounted } from "vue";
+import { useAuthentication } from "./ui/stores";
 import IndexView from "./ui/views/IndexView.vue";
 import HunterNavbar from "./ui/components/HunterNavbar.vue";
+
+const { verifySession } = useAuthentication();
+
+onMounted(async () => {
+  await verifySession();
+});
 </script>
 
 <template>